CLN: add shg_pl_trans module with relocated SHG-PL-Trans objects
