+++ pyfibre/shg_pl_trans/shg_pl_trans_factory.py
@@ -1,0 +1,20 @@
+from pyfibre.model.core.base_multi_image_factory import BaseMultiImageFactory
+
+from .shg_pl_trans_analyser import SHGPLTransAnalyser
+from .shg_pl_reader import SHGPLTransReader
+
+
+class SHGPLTransFactory(BaseMultiImageFactory):
+
+    def get_tag(self):
+        return 'SHG-PL-Trans'
+
+    def get_reader(self):
+        """Returns list of BaseMultiImageReader classes able to load
+        the BaseMultiImage class created by this factory"""
+        return SHGPLTransReader
+
+    def get_analyser(self):
+        """Returns list of BaseMultiImageAnalyser classes able to analyse
+        the BaseMultiImage class created by this factory"""
+        return SHGPLTransAnalyser
