+++ pyfibre/shg_pl_trans/tests/test_shg_analyser.py
@@ -3,7 +3,7 @@ from tempfile import TemporaryDirectory
 
 from pandas import DataFrame, Series
 
-from pyfibre.model.analysers.shg_analyser import SHGAnalyser
+from pyfibre.shg_pl_trans_plugin.shg_analyser import SHGAnalyser
 from pyfibre.pyfibre_runner import PyFibreRunner
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
 from pyfibre.tests.probe_classes.objects import (
