+++ pyfibre/shg_pl_trans/tests/test_shg_pl_trans_analyser.py
@@ -4,7 +4,7 @@ from tempfile import TemporaryDirectory
 import numpy as np
 from pandas import DataFrame, Series
 
-from pyfibre.model.analysers.shg_pl_trans_analyser import (
+from pyfibre.shg_pl_trans_plugin.shg_pl_trans_analyser import (
     SHGPLTransAnalyser)
 from pyfibre.pyfibre_runner import PyFibreRunner
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
