+++ pyfibre/shg_pl_trans/tools/tests/test_segmentation.py
@@ -1,8 +1,7 @@
 from skimage.io import imread
 
-from pyfibre.model.tools.segmentation import (
-    rgb_segmentation
-)
+from pyfibre.shg_pl_trans_plugin.tools.segmentation import (
+    shg_segmentation, shg_pl_trans_segmentation)
 from pyfibre.tests.probe_classes.utilities import (
     generate_image, generate_probe_graph)
 from pyfibre.tests.probe_classes.objects import ProbeFibreNetwork
@@ -25,13 +24,17 @@ class TestSegmentation(PyFibreTestCase):
         self.multi_image = ProbeSHGPLTransImage()
         self.fibre_networks = [ProbeFibreNetwork()]
 
-    def test_rgb_segmentation(self):
+    def test_shg_segmentation(self):
 
-        stack = (self.image_stack[0],
-                 self.image_stack[1],
-                 self.image_stack[2])
+        fibre_segments, cell_segments = shg_segmentation(
+            self.multi_image, self.fibre_networks
+        )
 
-        fibre_mask, cell_mask = rgb_segmentation(stack)
+        self.assertEqual(0, len(fibre_segments))
+        self.assertEqual(1, len(cell_segments))
 
-        self.assertEqual((200, 200), fibre_mask.shape)
-        self.assertEqual((200, 200), cell_mask.shape)
+    def test_shg_pl_trans_segmentation(self):
+
+        shg_pl_trans_segmentation(
+            self.multi_image, self.fibre_networks
+        )
